backend/socify-main/requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
openai
Werkzeug
gunicorn
python-dotenv
pillow
flask-cors